Re-organize solver and time_scheme
